comitting so I can merge in upstream
Former-commit-id: 32ceea4e2477b1f6d4549695449317404e199c4c
Former-commit-id: 8c7b75873c1b776cba87b6354e5b0f2a56f270c3

+++ NINA/Model/MyFlatDevice/FlatDeviceInfo.cs
@@ -15,9 +15,9 @@
             set { _lightOn = value; RaisePropertyChanged(); }
         }
 
-        private int _brightness;
+        private double _brightness;
 
-        public int Brightness {
+        public double Brightness {
             get => _brightness;
             set { _brightness = value; RaisePropertyChanged(); }
         }
